backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
google-genai==1.56.0
python-multipart==0.0.12
python-dotenv==1.0.1

